requirements.txt
numpy
pyyaml
python-dotenv
pillow
IPython
shapely
moviepy==1.0.3
networkx
cairosvg
matplotlib
transformers
diffusers
accelerate
jaxtyping
imageio[pyav]
scikit-image
omegaconf
opencv-python
BeautifulSoup4
visdom
# triton
numba
scipy
timm
scikit-fmm
einops
safetensors
datasets
scikit-learn
kornia
ftfy
regex
tqdm
rasterio
clip @ git+https://github.com/openai/CLIP.git
image-reward
google-genai

